Добавлена библиотека ProtocolLibrary, куда перенесена вся логика создания и работы SineGenerator и PackageParser. Исправлены ошибки, связанные с названиями полей и методов. Исправлены ошибки, связанные с утечкой памяти.

+++ ProtocolLibrary/src/PackageParserForInt16.cpp
@@ -14,7 +14,7 @@ void PackageParserForInt16::parsePackage(QByteArray data) {
 
     QString str;
     for (int i = 0; i < count; ++i) {
-        str += QString::number(float(sinus[i]) / float(std::numeric_limits<qint32>::max())) + ", ";
+        str += QString::number(float(sinus[i]) / float(std::numeric_limits<qint16>::max())) + ", ";
     }
 
     qDebug() << "Server received:" << str;
